Move library projections onto their produced types
The pure `Db*` → UI projections lived as free functions in
`manager/resolve.rs`, conflated with the I/O orchestration that reads the
database and coven cache. A projection is a property of the type it produces,
so each resolved type now owns its constructor: `ReleaseSummary::from_raw`,
`ReleaseDetail::from_raw`, `AlbumSummary::from_raw`, `ReleaseStorageSummary`/
`StorageRow::from_raw`, `SearchResults::from_raw`, and a genuine
`From<DbTrackSearchResult>` for `TrackSearchResult`. The single-use helpers
fold into those constructors.

`album_detail` becomes a directory so each produced type sits beside its
constructor. The pile of release-resolve parameters (which carried a
`too_many_arguments` allow) collapses into a `ReleaseResolveCtx` holding the
three coven-read inputs the manager threads: cloud-home presence, pin state,
and the release's own cover.

Only the pure projections move. The I/O resolvers (`find_release_detail_with`,
`cover_ref_for`, `resolve_album_detail`) stay in `resolve.rs` and now build a
`ReleaseResolveCtx` before handing off to the constructors. No behavior change.
