+++ bae-core/src/album_detail/search.rs
@@ -1,0 +1,95 @@
+//! Resolved search-result types (`SearchResults`, `AlbumSearchResult`,
+//! `TrackSearchResult`) and the pure projections that produce them.
+
+use std::collections::HashMap;
+
+use super::*;
+use crate::db::{DbLibrarySearchResults, DbTrackSearchResult};
+
+/// Resolved search-result container produced by `LibraryManager` from
+/// `DbLibrarySearchResults`.
+#[derive(Debug, Clone)]
+pub struct SearchResults {
+    pub albums: Vec<AlbumSearchResult>,
+    pub tracks: Vec<TrackSearchResult>,
+}
+
+impl SearchResults {
+    /// Resolve the raw search-result container by mapping each inner list.
+    /// `covers` maps an album's primary release id to its cover reference.
+    pub(crate) fn from_raw(
+        raw: DbLibrarySearchResults,
+        covers: &HashMap<String, ImageRef>,
+    ) -> SearchResults {
+        SearchResults {
+            albums: raw
+                .albums
+                .into_iter()
+                .map(|raw| {
+                    // The primary release id resolves the cover but isn't surfaced —
+                    // the search UI navigates by album id. The raw `primary_release_id`
+                    // comes from SQL's `COALESCE(a.primary_release_id, <first release
+                    // id>)` and is non-null by construction: every album has at least
+                    // one release (enforced by `delete_release`).
+                    let primary_release_id = raw
+                        .primary_release_id
+                        .expect("album has at least one release");
+                    let cover = covers.get(&primary_release_id).cloned();
+                    AlbumSearchResult {
+                        id: raw.id,
+                        title: raw.title,
+                        year: raw.year,
+                        artist_name: raw.artist_name,
+                        cover,
+                    }
+                })
+                .collect(),
+            tracks: raw
+                .tracks
+                .into_iter()
+                .map(TrackSearchResult::from)
+                .collect(),
+        }
+    }
+}
+
+/// Resolved album search result. Field-by-field copy of
+/// `DbAlbumSearchResult` — no transformation, just the "public" name.
+#[derive(Debug, Clone)]
+pub struct AlbumSearchResult {
+    pub id: String,
+    pub title: String,
+    pub year: Option<i32>,
+    pub artist_name: String,
+    /// Reference to the album's cover — the primary release's cover (image id +
+    /// version) — or `None` when it has no cover row. The search UI fetches its
+    /// bytes by image id and caches under `(id, version)`. Resolved from the
+    /// album's primary release; the primary id itself isn't surfaced (the search
+    /// UI navigates by album id).
+    pub cover: Option<ImageRef>,
+}
+
+/// Resolved track search result, produced from `DbTrackSearchResult`.
+#[derive(Debug, Clone)]
+pub struct TrackSearchResult {
+    pub id: String,
+    pub title: String,
+    pub duration_ms: Option<i64>,
+    pub album_id: String,
+    pub album_title: String,
+    pub artist_name: String,
+}
+
+impl From<DbTrackSearchResult> for TrackSearchResult {
+    /// Field-by-field copy into the display-ready shape.
+    fn from(raw: DbTrackSearchResult) -> Self {
+        TrackSearchResult {
+            id: raw.id,
+            title: raw.title,
+            duration_ms: raw.duration_ms,
+            album_id: raw.album_id,
+            album_title: raw.album_title,
+            artist_name: raw.artist_name,
+        }
+    }
+}
